Ajout des class test + légère modification des class modele (medicament et offrir)

+++ src/gsb/modele/Medicament.java
@@ -6,11 +6,11 @@ public class Medicament {
 	protected String composition;
 	protected String effets;
 	protected String contreIndication;
-	protected float prixEchantillon;
+	protected double prixEchantillon;
 	protected String codeFamille;
 	protected String libelleFamille;
 	
-	public Medicament(String depotLegal, String nomCommercial, String composition, String effets, String contreIndication, float prixEchantillon, String codeFamille, String libelleFamille) {
+	public Medicament(String depotLegal, String nomCommercial, String composition, String effets, String contreIndication, double prixEchantillon, String codeFamille, String libelleFamille) {
 		this.depotLegal = depotLegal;
 		this.nomCommercial = nomCommercial;
 		this.composition = composition;
@@ -61,7 +61,7 @@ public class Medicament {
 		this.contreIndication = contreIndication;
 	}
 	
-	public float getPrixEchantillon() {
+	public double getPrixEchantillon() {
 		return prixEchantillon;
 	}
